==> bookdown/04_prognostic_model.Rmd <==
## 
Development and internal validation of a multivariable prognostic model to predict chronic pain after a new episode of non-specific idiopathic, non-traumatic neck pain in physiotherapy primary care practice. 

```{r, include = FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>",
  warning = FALSE,
  error = FALSE,
  message = FALSE
)
```

## Packages
```{r}
library(rms)
library(foreign)
library(dplyr)
library(car)
library(fpc)
library(lmtest)
library(pROC)
library(tidyverse)
library(haven)
library(corrr)
library(car)
library(stats)
library(base)
library(broom)
library(MASS)
library(pROC)
library(ResourceSelection)
library(xfun)
library(skimr)
library(DataExplorer)
```

## Data load
We load the dataset that contains imputed missing values. To see the details on imputation method this package contains a vignette called 'imputation'. To see this and other vignettes and code run:
```{r, eval=FALSE}
browseVignettes(package = "painr")
```

The dataset is loaded from folder `./data` as an .rds R binary file.
```{r}
imputed_with_attitude_and_chronic <- read_rds(
  here::here(
    "data",
    "imputed_with_attitude_and_chronic.rds"
  )
)
```

## Clean data and rename vars
We rename the variable `wide_spread_pain` to `pidbr` (pain in different body regions) 
```{r}
imputed_with_attitude_and_chronic <- imputed_with_attitude_and_chronic |>
  rename(pidbr = wide_spread_pain)
```

## Exploratory Data Analysis
With `summary()` we can get an overview of the different variables
```{r}
summary(imputed_with_attitude_and_chronic)
```

Using a more sophisticated workflow to do exploratory analysis can be done using the `{skimr}` package and the `{DataExplorer}` package. To get an automated report for this dataset, run the following code. The file can be found in the `./inst` folder of this repo.
```{r, eval=FALSE}
imputed_with_attitude_and_chronic |>
    create_report(
        output_file  = "eda_DataExplorer_imputed_with_attitude_and_chronic",
        output_dir   = "inst",
        y            = "painint_total_score",
        report_title = "EDA painint_total_score"
    )
```

## Variabel analysis - Independent Predictive Capacity 

We conducted univariable logistic regression analyses to assess the independent predictive capacity of the candidate prognostic variables at baseline and their association with the presence or absence of chronic pain at six weeks, 
three months, and six months. For continuous variables, we applied logistic regression, fitting a univariate logistic regression model with a 95% confidence interval (CI) and odds ratio (OR). 
Additionally, we extracted the coefficient and standard error for each variable.

We did not use these analyses to decide which prognostic factors will be included in the multivariable analyses.


```{r}
continuous_variables <-
  c(
    "depression",
    "concerns",
    "age",
    "pain_intensity",
    "duration",
    "disability",
    "bmi",
    "catastrophizing",
    "duration_beliefs",
    "treatment_beliefs",
    "kinesiophobia",
    "distress",
    "identity_beliefs",
    "hypervigilance",
    "relation",
    "self_efficacy"
  )

# Univariate analyses of the continuous variables
univariate_analysis_continuous <- function(variable, data) {
  formula <- as.formula(paste("is_painint_chronic ~", variable))
  model <- glm(formula, data = data, family = binomial)
  summary_model <- summary(model)
  coef_variable <- summary_model$coefficients[variable, "Estimate"]
  std_error <- summary_model$coefficients[variable, "Std. Error"]
  OR <- exp(coef_variable)
  lower_limit <- exp(coef_variable - 1.96 * std_error)
  upper_limit <- exp(coef_variable + 1.96 * std_error)
  cat("Variable:", variable, "\n")
  cat("Odds Ratio (OR):", OR, "\n")
  cat("95% Betrouwbaarheidsinterval:",
      lower_limit,
      "-",
      upper_limit,
      "\n\n")
}

# Univariate analyses for all continuous variables
for (variable in continuous_variables) {
  univariate_analysis_continuous(variable, imputed_with_attitude_and_chronic)
}

```

```{r}

# Variable "work"
imputed_with_attitude_and_chronic$work <-
  relevel(imputed_with_attitude_and_chronic$work, ref = "1")
model1 <-
  glm(is_painint_chronic ~ work, data = imputed_with_attitude_and_chronic, family = binomial)
OR_work <- exp(coef(model1))
CI_work <- exp(confint(model1))
print(OR_work)
print(CI_work)

# Variable "Education_level"
imputed_with_attitude_and_chronic$education_level <-
  relevel(imputed_with_attitude_and_chronic$education_level, ref = "1")
model2 <-
  glm(is_painint_chronic ~ education_level,
      data = imputed_with_attitude_and_chronic,
      family = binomial)
OR_education_level <- exp(coef(model2))
CI_education_level <- exp(confint(model2))
print(OR_education_level)
print(CI_education_level)

# Variable "recurrence"
imputed_with_attitude_and_chronic$recurrence <-
  relevel(imputed_with_attitude_and_chronic$recurrence, ref = "1")
model3 <-
  glm(is_painint_chronic ~ recurrence,
      data = imputed_with_attitude_and_chronic,
      family = binomial)
OR_recurrence <- exp(coef(model3))
CI_recurrence <- exp(confint(model3))
print(OR_recurrence)
print(CI_recurrence)

# Variable - pain in different body regions
imputed_with_attitude_and_chronic$pidbr <-
  relevel(imputed_with_attitude_and_chronic$pidbr, ref = "1")
model4 <-
  glm(is_painint_chronic ~ pidbr, data = imputed_with_attitude_and_chronic, family = binomial)
OR_pidbr <- exp(coef(model4))
CI_pidbr <- exp(confint(model4))
print(OR_pidbr)
print(CI_pidbr)

# Variable "headache"
imputed_with_attitude_and_chronic$headache <-
  relevel(imputed_with_attitude_and_chronic$headache, ref = "1")
model5 <-
  glm(is_painint_chronic ~ headache,
      data = imputed_with_attitude_and_chronic,
      family = binomial)
OR_headache <- exp(coef(model5))
CI_headache <- exp(confint(model5))
print(OR_headache)
print(CI_headache)

# Variable "sex"
imputed_with_attitude_and_chronic$sex <-
  relevel(imputed_with_attitude_and_chronic$sex, ref = "1")
model6 <-
  glm(is_painint_chronic ~ sex, data = imputed_with_attitude_and_chronic, family = binomial)
OR_sex <- exp(coef(model6))
CI_sex <- exp(confint(model6))
print(OR_sex)
print(CI_sex)

# Variable "Work_happiness"
imputed_with_attitude_and_chronic$work_happiness <-
  relevel(imputed_with_attitude_and_chronic$work_happiness, ref = "1")
model7 <-
  glm(is_painint_chronic ~ work_happiness,
      data = imputed_with_attitude_and_chronic,
      family = binomial)
OR_work_happiness <- exp(coef(model7))
CI_work_happiness <- exp(confint(model7))
print(OR_work_happiness)
print(CI_work_happiness)

# Variable "posture_work"
imputed_with_attitude_and_chronic$posture_work <-
  relevel(imputed_with_attitude_and_chronic$posture_work, ref = "1")
model8 <-
  glm(is_painint_chronic ~ posture_work,
      data = imputed_with_attitude_and_chronic,
      family = binomial)
OR_posture_work <- exp(coef(model8))
CI_posture_work <- exp(confint(model8))
print(OR_posture_work)
print(CI_posture_work)

# Variable "work_satisfaction"
imputed_with_attitude_and_chronic$work_satisfaction <-
  relevel(imputed_with_attitude_and_chronic$work_satisfaction, ref = "1")
model9 <-
  glm(is_painint_chronic ~ work_satisfaction,
      data = imputed_with_attitude_and_chronic,
      family = binomial)
OR_work_satisfaction <- exp(coef(model9))
CI_work_satisfaction <- exp(confint(model9))
print(OR_work_satisfaction)
print(CI_work_satisfaction)

# Variable "physical Activity"
imputed_with_attitude_and_chronic$physical_activity <-
  relevel(imputed_with_attitude_and_chronic$physical_activity, ref = "0")
model10 <-
  glm(is_painint_chronic ~ physical_activity,
      data = imputed_with_attitude_and_chronic,
      family = binomial)
OR_physical_activity <- exp(coef(model10))
CI_physical_activity <- exp(confint(model10))
print(OR_physical_activity)
print(CI_physical_activity)

# Variable "smoking"
imputed_with_attitude_and_chronic$smoking <-
  relevel(imputed_with_attitude_and_chronic$smoking, ref = "1")
model11 <-
  glm(is_painint_chronic ~ smoking,
      data = imputed_with_attitude_and_chronic,
      family = binomial)
OR_smoking <- exp(coef(model11))
CI_smoking <- exp(confint(model11))
print(OR_smoking)
print(CI_smoking)

# Variable "alcohol"
imputed_with_attitude_and_chronic$alcohol <-
  relevel(imputed_with_attitude_and_chronic$alcohol, ref = "1")
model12 <-
  glm(is_painint_chronic ~ alcohol,
      data = imputed_with_attitude_and_chronic,
      family = binomial)
OR_alcohol <- exp(coef(model12))
CI_alcohol <- exp(confint(model12))
print(OR_alcohol)
print(CI_alcohol)

# Variable "sleep quality"
imputed_with_attitude_and_chronic$sleep_quality <-
  relevel(imputed_with_attitude_and_chronic$sleep_quality, ref = "0")
model13 <-
  glm(is_painint_chronic ~ sleep_quality,
      data = imputed_with_attitude_and_chronic,
      family = binomial)
OR_sleep_quality <- exp(coef(model13))
CI_sleep_quality <- exp(confint(model13))
print(OR_sleep_quality)
print(CI_sleep_quality)

# Variable "coping"
imputed_with_attitude_and_chronic$coping <-
  relevel(imputed_with_attitude_and_chronic$coping, ref = "0")
model14 <-
  glm(is_painint_chronic ~ coping,
      data = imputed_with_attitude_and_chronic,
      family = binomial)
OR_coping <- exp(coef(model14))
CI_coping <- exp(confint(model14))
print(OR_coping)
print(CI_coping)

# Variable "attitude"
model15 <-
  glm(is_painint_chronic ~ attitude,
      data = imputed_with_attitude_and_chronic,
      family = binomial)
OR_attitude <- exp(coef(model15))
CI_attitude <- exp(confint(model15))
print(OR_attitude)
print(CI_attitude)
```

## Multivariable logistic regression analyses

Due to multicollinearity among the work-related factors (happiness, satisfaction, and posture, leading to the outcome '3' = not working), we had to make a decision regarding which factor to include. 
We analyzed different models with each of the individual work-related factors included and found that there was almost no difference in model performance. 
Since happiness and satisfaction align more closely with the other psychological factors, we chose to include the variable 'posture_work,' thereby incorporating a different domain into our 
model for the final backward model analyses and internal validation.

We found that the candidate prognostic factor 'attitude' did not have predictive value in the univariate analysis and did not emerge in the multivariate analysis when all variables were included. 
Additionally, there were many missing values that could not be imputed, as they pertain to the therapist's attitude. 
Therefore, we decided to exclude this variable from the complete model development, as internal validation cannot proceed with these variables included.

```{r}
naniar::vis_miss(imputed_with_attitude_and_chronic)
new_data <- imputed_with_attitude_and_chronic |>
  dplyr::select(-attitude)

sum(is.na(new_data))

full_model <-
  glm (
    data = imputed_with_attitude_and_chronic,
    is_painint_chronic ~ sex + age + pain_intensity + duration + pidbr + headache + disability + posture_work + physical_activity + smoking + alcohol + bmi + sleep_quality + catastrophizing + duration_beliefs + concerns + treatment_beliefs + depression + kinesiophobia + distress + coping + identity_beliefs + hypervigilance + self_efficacy + relation,
    family = "binomial"
  )

summary(full_model)

#full_model_mterics_df <- full_model |> broom::tidy()
#full_model_mterics_df <- full_model_mterics_df |>
#  mutate(
#    OR = exp(coef(estimate))
#  )

backward_model <- stepAIC(full_model, direction = "backward")
summary(backward_model)
backward_model |> broom::tidy()

odds_ratios <- exp(coef(backward_model))
conf_int <- confint(backward_model)
conf_int_exp <- exp(conf_int)

results <- data.frame(OddsRatio = odds_ratios,
                      Lower95CI = conf_int_exp[, "2.5 %"],
                      Upper95CI = conf_int_exp[, "97.5 %"])
print(results)

```

## AUC
We now will determine the discriminative ability of the prognostic model based on the Area Under the Receiver Operating Characteristic Curve (AUC).

```{r}
# Prediction of the probabilities based on our model
predicted_probs <-
  predict(backward_model, newdata = imputed_with_attitude_and_chronic, type =
            "response")

roc_obj <-
  roc(imputed_with_attitude_and_chronic$is_painint_chronic,
      predicted_probs)
auc <- auc(roc_obj)
print(auc)

ci <- ci.auc(roc_obj)
print(ci)

```


## Calibration curve
We will measure the predicted probabilities and the observed responses, create a dataframe, group the predicted probabilities into deciles, 
calculate the average predicted probability and actual percentage for each group, and plot the calibration curve


```{r} 
model_formula <-
  is_painint_chronic ~ sex + pain_intensity + pidbr + headache + posture_work + 
  duration_beliefs + concerns + treatment_beliefs + distress + identity_beliefs + 
  self_efficacy

# Model fitting in lrm.
backward_model_lrm <-
  lrm(model_formula,
      data = imputed_with_attitude_and_chronic,
      x = TRUE,
      y = TRUE)

predicted_probs_lrm <- predict(backward_model_lrm, type = "fitted")

val.prob(
  predicted_probs_lrm,
  as.numeric(imputed_with_attitude_and_chronic$is_painint_chronic),
  statloc = FALSE,
  logistic.cal = FALSE
)

calibration_plot_lrm <-
  calibrate(backward_model_lrm, method = "boot", B = 200)

plot(calibration_plot_lrm)

```

## Result Calibration plot 
The calibration plot revealed acceptable alignment for patients not developing chronic pain. 
However, for patients developing chronic pain, the model tended to overestimate the risk. 
Despite this, the calibration remained within acceptable limits.

## Hoslem and Lemeshow 
We will formally test the goodness-of-fit using the Hosmer and Lemeshow test.

```{r}
predicted_probs <-
  predict(
    backward_model, 
    newdata = imputed_with_attitude_and_chronic, 
    type = "response")

gof <-
  hoslem.test(
    imputed_with_attitude_and_chronic$is_painint_chronic,
    predicted_probs,
    g = 10)

gof |> broom::tidy()

x_sqrd <- gof$statistic[1]
p_value_gof <- gof$p.value[1]
```

Results: 
X-squared = `r x_sqrd`, DF = 8, p-value = `r p_value_gof` 
We found a high p-value suggests there's no statistically significant difference between observed and expected frequencies, indicating = that the model fits the data well.


## Model fit 
Will be quantified as Nagelkerke’s R2

```{r}
nagelkerkeR2 <- function(model) {
  LL_0 <- as.numeric(logLik(update(model, .~1)))
  LL_f <- as.numeric(logLik(model))
  R2 <- (1 - exp((2/nrow(model$data))*(LL_0 - LL_f))) /
    (1 - exp(2*LL_0/nrow(model$data)))
  return(R2)
}

nagelkerkeR2(backward_model)
```


## Internal validation 
Internal validation will be performed using bootstrap resampling to estimate the optimism-corrected AUC and to yield a measure of overfitting (i.e., the shrinkage factor). 
The shrinkage factor (a constant between 0 and 1) will be used to multiply the regression coefficient by. 
Generally, regression coefficients(and resulting predictions) are too extreme in case of overfitting, which is counteracted by the shrinking of regression coefficients.

```{r}


set.seed(12345)

formula(backward_model)

new_model <-
  lrm(formula(backward_model),
      data = imputed_with_attitude_and_chronic,
      x = TRUE,
      y = TRUE)

validation_result <- validate(new_model, method = "boot", B = 1000)

print(validation_result)

original_dxy <- validation_result["Dxy", "index.orig"]
optimism_dxy <- validation_result["Dxy", "optimism"]

corrected_dxy <- original_dxy - optimism_dxy
print(corrected_dxy)

original_auc <- original_dxy / 2 + 0.5
optimism_auc <- optimism_dxy / 2

corrected_auc <- original_auc - optimism_auc
print(corrected_auc)

```

## Plot Corrected AUC

```{r}
roc_obj <-
  roc(response = imputed_with_attitude_and_chronic$is_painint_chronic,
      predictor = predicted_probs)

plot(
  roc_obj$specificities,
  roc_obj$sensitivities,
  type = "l",
  xlim = c(1, 0),
  xlab = "1 - Specificity",
  ylab = "Sensitivity",
  main = "ROC Curve"
)

```
## Results
Corrected AUC = 0.8264617
Our original AUC was 0.86 on the entire dataset. 
The corrected AUC is not much lower, indicating that overfitting is relatively low. 
Therefore, our model appears to have a robust discriminatory ability



## Correcting the variables coefficients
We will multiply all the coefficients of our model by the shrinkage factor 'sf' and estimate the intercept so that the average estimated probability is equal to the frequency of the outcome.

```{r}
predictiemodel <- round(new_model$coef * 0.8274, 3)
predictiemodel.lp <- cbind(new_model$x) %*% predictiemodel[-1]

predictiemodel["Intercept"] <-
  round(
    lrm.fit(y = imputed_with_attitude_and_chronic$is_painint_chronic, offset =
              predictiemodel.lp)$coef,
    3
  )

print (predictiemodel)

```

## Calibration in the Large 
We will check whether the internally validated model provides the same average estimated probability as the average outcome (calibration in the large). 
We will calculate the linear predictor for everyone in the database, create a function to calculate the risk using the formula 1/(1+exp(-(LP))),
and then calculate the probability for each individual in the database and the average probability.

```{r}
predictiemodel.linearpredictor <-
  predictiemodel[1] + new_model$x %*% predictiemodel[-1]

risk <- function(lp)
  
{
  risk <- 1 / (1 + exp(-lp))
}

risk.predicted <- risk(predictiemodel.linearpredictor)

mean(risk.predicted)

mean(imputed_with_attitude_and_chronic$is_painint_chronic)

```

# Result = 0.1028 and 0.1028
The model is well calibrated in the large.
